+++ src/crypto/ctr/CTRCrypto.ts
@@ -3,6 +3,7 @@ import * as crypto from 'crypto';
 
 export class CTRCrypto implements BaseCryptoApi {
 
+    private iv  = Buffer.from('26ae5cc854e36b6bdfca366848dea6bb', 'hex');
     private cryptoData : any = {};
     constructor(data : any) {
         this.cryptoData.algorithm = 'aes-256-ctr';
@@ -13,14 +14,14 @@ export class CTRCrypto implements BaseCryptoApi {
     }
 
     encrypt(text : string) {
-        var cipher = crypto.createCipher(this.cryptoData.algorithm, this.cryptoData.password)
+        var cipher = crypto.createCipheriv(this.cryptoData.algorithm, this.cryptoData.password, this.iv)
         var crypted = cipher.update(text, 'utf8', 'hex')
         crypted += cipher.final('hex');
         return crypted;
     }
 
     decrypt(encrypted : string) {
-        var decipher = crypto.createDecipher(this.cryptoData.algorithm, this.cryptoData.password)
+        var decipher = crypto.createDecipheriv(this.cryptoData.algorithm, this.cryptoData.password, this.iv)
         var dec = decipher.update(encrypted, 'hex', 'utf8')
         dec += decipher.final('utf8');
         return dec;